+++ tests/audio.rs
@@ -121,6 +121,13 @@ fn test_mpeg12_audio_layer3() {
     assert_eq!(fmt, FileFormat::Mpeg12AudioLayer3);
 }
 
+#[cfg(feature = "reader-mp4")]
+#[test]
+fn test_mpeg4_part14_audio() {
+    let fmt = FileFormat::from_file("fixtures/audio/sample.mp4").unwrap();
+    assert_eq!(fmt, FileFormat::Mpeg4Part14Audio);
+}
+
 #[test]
 fn test_musepack() {
     let fmt = FileFormat::from_file("fixtures/audio/sample.mpc").unwrap();