+++ tests/video.rs
@@ -105,6 +105,7 @@ fn test_mpeg2_transport_stream() {
     assert_eq!(fmt, FileFormat::Mpeg2TransportStream);
 }
 
+#[cfg(feature = "reader-mp4")]
 #[test]
 fn test_mpeg4_part14_video() {
     let fmt = FileFormat::from_file("fixtures/video/sample.mp4").unwrap();