+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Version 0.21.0 (unreleased)
+# Version 0.21.0 (2023-09-29)
 
 ## API